Revert "Harpoon release 2.0.0 for MCUX_2.10.0"
This reverts commit 46d44c68e6cb7f97a68e6a1f75aaf4337dba90b5.

+++ components/codec/port/wm8524/fsl_codec_adapter.c
@@ -38,7 +38,7 @@ status_t HAL_CODEC_Init(void *handle, void *config)
     codec_config_t *codecConfig = (codec_config_t *)config;
 
     wm8524_config_t *devConfig = (wm8524_config_t *)(codecConfig->codecDevConfig);
-    wm8524_handle_t *devHandle = (wm8524_handle_t *)((uintptr_t)(((codec_handle_t *)handle)->codecDevHandle));
+    wm8524_handle_t *devHandle = (wm8524_handle_t *)((uint32_t)(((codec_handle_t *)handle)->codecDevHandle));
 
     ((codec_handle_t *)handle)->codecCapability = &s_wm8524_capability;
 
@@ -102,7 +102,7 @@ status_t HAL_CODEC_SetMute(void *handle, uint32_t playChannel, bool isMute)
 {
     assert(handle != NULL);
 
-    WM8524_SetMute((wm8524_handle_t *)((uintptr_t)(((codec_handle_t *)handle)->codecDevHandle)), isMute);
+    WM8524_SetMute((wm8524_handle_t *)((uint32_t)(((codec_handle_t *)handle)->codecDevHandle)), isMute);
 
     return kStatus_Success;
 }
